Bugfix: int should be long

+++ src/net/sroz/grocerylist/GroceryItem.java
@@ -3,12 +3,12 @@ package net.sroz.grocerylist;
 import android.database.Cursor;
 
 public class GroceryItem {
-	public int id;
+	public long id;
 	public String text;
 	public boolean checked;
 		
 	public GroceryItem(Cursor c) {
-		id = c.getInt(c.getColumnIndex(GroceryProvider.KEY_ROWID));
+		id = c.getLong(c.getColumnIndex(GroceryProvider.KEY_ROWID));
 		text = c.getString(c.getColumnIndex(GroceryProvider.KEY_TEXT));
 		checked = (c.getInt(c.getColumnIndex(GroceryProvider.KEY_CHECKED)) == 1);
 	}
